+++ tests/testthat/test-bm_glyph.R
@@ -3,7 +3,7 @@ test_that("bm_glyph()", {
     expect_error(bm_glyph("Zippity"))
     expect_error(bm_glyph(matrix("Zippity")))
 
-    space_matrix <- matrix(0, nrow = 16, ncol = 16)
+    space_matrix <- matrix(0L, nrow = 16, ncol = 16)
     space_glyph <- bm_glyph(space_matrix)
     expect_true(is_bm_glyph(space_glyph))
     expect_false(is_bm_glyph(space_matrix))
@@ -15,4 +15,7 @@ test_that("bm_glyph()", {
 
     space_glyph3 <- as_bm_glyph(space_matrix)
     expect_equal(space_glyph, space_glyph3)
+
+    space_matrix2 <- as.matrix(space_glyph)
+    expect_equal(space_matrix, space_matrix2)
 })